requirements.txt
Django==5.0.7
redis==5.0.8
djangorestframework==3.15.2
